Missing value gap statistics

Computes the average missing value gap of a vector.

gapMiss(x, what = mean)

Arguments

x

a numeric vector

what

default is the arithmetic mean. One can include an own function that returns a vector of lenght 1 (e.g. median)

Value

The gap statistics

Details

The length of each sequence of missing values (gap) in a vector is calculated and the mean gap is reported

Author

Matthias Templ based on a suggestion and draft from Huang Tian Yuan.

Examples

v <- rnorm(20)
v[3] <- NA
v[6:9] <- NA
v[13:17] <- NA
v
#>  [1] -1.1028021  0.4897198         NA  0.4677914  1.5351436         NA
#>  [7]         NA         NA         NA  1.1995744  1.1842430 -0.6070301
#> [13]         NA         NA         NA         NA         NA  0.3959722
#> [19]  0.2859305  1.4599684
gapMiss(v)
#> [1] 3.333333
gapMiss(v, what = median)
#> [1] 4
gapMiss(v, what = function(x) mean(x, trim = 0.1))
#> [1] 3.333333
gapMiss(v, what = var)
#> [1] 4.333333