Synthetic Austrian Structural Business Survey data

A synthetic, non-confidential enterprise-level data set inspired by the Austrian Leistungs- und Strukturstatistik from reporting years 2021 onward.

Format

A data frame with 10000 observations and 38 variables.

Source

Statistik Austria, Standard-Dokumentation zur Leistungs- und Strukturstatistik ab 2021, https://www.statistik.at/fileadmin/shared/QM/Standarddokumentationen/U/std_u_ls_ab_2021.pdf

Details

The variables cover the main dimensions described in the Statistik Austria standard documentation: economic activity, NUTS2 region, data source, employment structure, turnover, intermediate consumption, gross value added, personnel costs, gross operating surplus, and tangible investment.

The monetary variables are expressed in thousand euros. The observations are generated synthetically and do not contain real enterprise records. Companion validation rules are available as lse_synthetic_rules.

Examples

data(lse_synthetic)
str(lse_synthetic)
#> 'data.frame':	10000 obs. of  38 variables:
#>  $ unit_id                 : chr  "LU00001" "LU00002" "LU00003" "LU00004" ...
#>  $ reporting_year          : int  2023 2022 2023 2022 2023 2023 2022 2022 2023 2021 ...
#>  $ enterprise_group_id     : chr  "ENT02148" "ENT02396" "ENT02323" "ENT04528" ...
#>  $ onace_section           : Factor w/ 17 levels "B","C","D","E",..: 2 3 3 16 4 13 6 5 7 13 ...
#>  $ onace_group             : Factor w/ 18 levels "B08","C10","C25",..: 9 14 13 9 18 5 10 11 15 6 ...
#>  $ nuts2                   : Factor w/ 9 levels "AT11","AT12",..: 5 6 5 8 7 4 2 2 4 8 ...
#>  $ data_source             : Factor w/ 3 levels "primary_survey",..: 2 1 3 3 2 3 2 2 3 2 ...
#>  $ survey_mode             : Factor w/ 4 levels "eQuest","lse_balance_list",..: 4 1 4 4 4 4 4 4 4 4 ...
#>  $ persons_employed        : num  9 23 22 19 71 11 19 79 18 12 ...
#>  $ employees_paid          : num  8 23 20 19 71 11 19 79 17 11 ...
#>  $ self_employed           : num  1 0 2 0 0 0 0 0 1 1 ...
#>  $ employees_male          : num  1 16 16 17 41 6 14 35 6 6 ...
#>  $ employees_female        : num  7 7 4 2 30 5 5 44 11 5 ...
#>  $ employees_blue_collar   : num  4 9 9 6 29 3 9 32 6 4 ...
#>  $ employees_white_collar  : num  1 11 10 12 37 7 7 33 8 6 ...
#>  $ apprentices             : num  0 1 0 0 4 0 2 2 1 0 ...
#>  $ marginal_employees      : num  3 2 1 1 1 1 1 12 2 1 ...
#>  $ employment_size_class   : Factor w/ 6 levels "0-1","2-9","10-19",..: 2 4 4 3 5 3 3 5 3 3 ...
#>  $ turnover_total          : num  1461 2501 1789 1972 4248 ...
#>  $ turnover_domestic       : num  1125 1876 1258 1459 3945 ...
#>  $ turnover_exports        : num  336 625 531 513 303 ...
#>  $ e_commerce_turnover     : num  294 479 499 110 997 382 687 991 568 17 ...
#>  $ turnover_size_class     : Factor w/ 6 levels "10-99","100-999",..: 3 4 3 3 4 3 4 4 4 3 ...
#>  $ material_costs          : num  484 255 369 369 1348 ...
#>  $ purchased_services      : num  154 222 115 157 329 ...
#>  $ rents_leasing           : num  32 67 80 30 292 74 76 391 57 78 ...
#>  $ other_operating_expense : num  123 178 126 179 143 34 222 447 192 119 ...
#>  $ intermediate_consumption: num  793 722 690 735 2112 ...
#>  $ gross_value_added       : num  668 1779 1099 1237 2136 ...
#>  $ personnel_costs         : num  357 893 503 718 2136 ...
#>  $ wages_salaries          : num  257 655 384 563 1474 ...
#>  $ social_security_costs   : num  66 179 116 115 511 58 150 599 77 139 ...
#>  $ other_personnel_costs   : num  34 59 3 40 151 21 71 0 15 18 ...
#>  $ gross_operating_surplus : num  311 886 596 519 0 386 528 0 408 398 ...
#>  $ investments_tangible    : num  146 184 184 60 242 89 372 123 286 110 ...
#>  $ investment_machinery    : num  58 74 109 35 130 21 76 64 132 65 ...
#>  $ investment_buildings    : num  31 32 25 0 60 10 41 0 60 7 ...
#>  $ investment_software     : num  57 78 50 25 52 58 255 59 94 38 ...