C-horizon of the Kola data with missing values

This data set is the same as the mvoutlier::chorizon() data set in package mvoutlier, except that values below the detection limit are coded as NA.

Format

A data frame with 606 observations on the following 110 variables.

*ID

a numeric vector

XCOO

a numeric vector

YCOO

a numeric vector

Ag

a numeric vector

Ag_INAA

a numeric vector

Al

a numeric vector

Al2O3

a numeric vector

As

a numeric vector

As_INAA

a numeric vector

Au_INAA

a numeric vector

B

a numeric vector

Ba

a numeric vector

Ba_INAA

a numeric vector

Be

a numeric vector

Bi

a numeric vector

Br_IC

a numeric vector

Br_INAA

a numeric vector

Ca

a numeric vector

Ca_INAA

a numeric vector

CaO

a numeric vector

Cd

a numeric vector

Ce_INAA

a numeric vector

Cl_IC

a numeric vector

Co

a numeric vector

Co_INAA

a numeric vector

EC

a numeric vector

Cr

a numeric vector

Cr_INAA

a numeric vector

Cs_INAA

a numeric vector

Cu

a numeric vector

Eu_INAA

a numeric vector

F_IC

a numeric vector

Fe

a numeric vector

Fe_INAA

a numeric vector

Fe2O3

a numeric vector

Hf_INAA

a numeric vector

Hg

a numeric vector

Hg_INAA

a numeric vector

Ir_INAA

a numeric vector

K

a numeric vector

K2O

a numeric vector

La

a numeric vector

La_INAA

a numeric vector

Li

a numeric vector

LOI

a numeric vector

Lu_INAA

a numeric vector

wt_INAA

a numeric vector

Mg

a numeric vector

MgO

a numeric vector

Mn

a numeric vector

MnO

a numeric vector

Mo

a numeric vector

Mo_INAA

a numeric vector

Na

a numeric vector

Na_INAA

a numeric vector

Na2O

a numeric vector

Nd_INAA

a numeric vector

Ni

a numeric vector

Ni_INAA

a numeric vector

NO3_IC

a numeric vector

P

a numeric vector

P2O5

a numeric vector

Pb

a numeric vector

pH

a numeric vector

PO4_IC

a numeric vector

Rb

a numeric vector

S

a numeric vector

Sb

a numeric vector

Sb_INAA

a numeric vector

Sc

a numeric vector

Sc_INAA

a numeric vector

Se

a numeric vector

Se_INAA

a numeric vector

Si

a numeric vector

SiO2

a numeric vector

Sm_INAA

a numeric vector

Sn_INAA

a numeric vector

SO4_IC

a numeric vector

Sr

a numeric vector

Sr_INAA

a numeric vector

SUM_XRF

a numeric vector

Ta_INAA

a numeric vector

Tb_INAA

a numeric vector

Te

a numeric vector

Th

a numeric vector

Th_INAA

a numeric vector

Ti

a numeric vector

TiO2

a numeric vector

U_INAA

a numeric vector

V

a numeric vector

W_INAA

a numeric vector

Y

a numeric vector

Yb_INAA

a numeric vector

Zn

a numeric vector

Zn_INAA

a numeric vector

ELEV

a numeric vector

*COUN

a numeric vector

*ASP

a numeric vector

TOPC

a numeric vector

LITO

a numeric vector

Al_XRF

a numeric vector

Ca_XRF

a numeric vector

Fe_XRF

a numeric vector

K_XRF

a numeric vector

Mg_XRF

a numeric vector

Mn_XRF

a numeric vector

Na_XRF

a numeric vector

P_XRF

a numeric vector

Si_XRF

a numeric vector

Ti_XRF

a numeric vector

Source

Kola Project (1993-1998)

Note

For a more detailed description of this data set, see mvoutlier::chorizon() in package mvoutlier.

References

Reimann, C., Filzmoser, P., Garrett, R.G. and Dutter, R. (2008) Statistical Data Analysis Explained: Applied Environmental Statistics with R. Wiley.

See also

Examples

data(chorizonDL, package = "VIM")
summary(chorizonDL)

#>       *ID             XCOO             YCOO               Ag         
#>  Min.   :  1.0   Min.   :372602   Min.   :7373050   Min.   :0.00100  
#>  1st Qu.:185.2   1st Qu.:498838   1st Qu.:7512490   1st Qu.:0.00500  
#>  Median :383.0   Median :601950   Median :7612020   Median :0.00800  
#>  Mean   :386.6   Mean   :603030   Mean   :7610244   Mean   :0.01095  
#>  3rd Qu.:578.8   3rd Qu.:701207   3rd Qu.:7701630   3rd Qu.:0.01300  
#>  Max.   :905.0   Max.   :861309   Max.   :7883400   Max.   :0.11900  
#>                                                     NA's   :1        
#>     Ag_INAA          Al            Al2O3             As        
#>  Min.   : NA   Min.   : 1840   Min.   : 5.52   Min.   : 0.100  
#>  1st Qu.: NA   1st Qu.: 6475   1st Qu.:13.06   1st Qu.: 0.300  
#>  Median : NA   Median : 9910   Median :13.95   Median : 0.600  
#>  Mean   :NaN   Mean   :12665   Mean   :13.88   Mean   : 1.271  
#>  3rd Qu.: NA   3rd Qu.:15500   3rd Qu.:14.79   3rd Qu.: 1.200  
#>  Max.   : NA   Max.   :85900   Max.   :22.83   Max.   :30.700  
#>  NA's   :606                                   NA's   :10      
#>     As_INAA          Au_INAA             B               Ba         
#>  Min.   : 0.250   Min.   :0.0020   Min.   : 3.00   Min.   :   4.70  
#>  1st Qu.: 0.250   1st Qu.:0.0030   1st Qu.: 3.00   1st Qu.:  26.73  
#>  Median : 0.250   Median :0.0040   Median : 3.00   Median :  43.50  
#>  Mean   : 1.362   Mean   :0.0109   Mean   : 3.75   Mean   :  60.15  
#>  3rd Qu.: 1.400   3rd Qu.:0.0060   3rd Qu.: 4.00   3rd Qu.:  71.90  
#>  Max.   :54.000   Max.   :0.9810   Max.   :11.00   Max.   :1300.00  
#>                   NA's   :439      NA's   :542                      
#>     Ba_INAA             Be                Bi              Br_IC    
#>  Min.   : 210.0   Min.   : 0.0600   Min.   :0.00600   Min.   : NA  
#>  1st Qu.: 460.0   1st Qu.: 0.1500   1st Qu.:0.01500   1st Qu.: NA  
#>  Median : 575.0   Median : 0.2350   Median :0.02700   Median : NA  
#>  Mean   : 599.6   Mean   : 0.4422   Mean   :0.05023   Mean   :NaN  
#>  3rd Qu.: 700.0   3rd Qu.: 0.3700   3rd Qu.:0.04900   3rd Qu.: NA  
#>  Max.   :3000.0   Max.   :14.0000   Max.   :3.89000   Max.   : NA  
#>                                     NA's   :15        NA's   :606  
#>     Br_INAA             Ca           Ca_INAA           CaO       
#>  Min.   : 1.100   Min.   :  110   Min.   :10000   Min.   :0.050  
#>  1st Qu.: 3.375   1st Qu.: 1272   1st Qu.:20000   1st Qu.:2.252  
#>  Median : 4.700   Median : 1905   Median :30000   Median :3.050  
#>  Mean   : 6.291   Mean   : 2279   Mean   :26688   Mean   :2.991  
#>  3rd Qu.: 7.200   3rd Qu.: 2748   3rd Qu.:30000   3rd Qu.:3.770  
#>  Max.   :56.000   Max.   :41700   Max.   :60000   Max.   :9.460  
#>  NA's   :154                      NA's   :138                    
#>        Cd             Ce_INAA           Cl_IC               Co       
#>  Min.   :0.00700   Min.   : 12.00   Min.   :0.00100   Min.   : 1.20  
#>  1st Qu.:0.01800   1st Qu.: 32.00   1st Qu.:0.00100   1st Qu.: 4.90  
#>  Median :0.02400   Median : 45.00   Median :0.00100   Median : 7.00  
#>  Mean   :0.02869   Mean   : 58.84   Mean   :0.00194   Mean   : 8.22  
#>  3rd Qu.:0.03300   3rd Qu.: 69.00   3rd Qu.:0.00300   3rd Qu.:10.10  
#>  Max.   :0.22100   Max.   :500.00   Max.   :0.01000   Max.   :44.30  
#>                                                                      
#>     Co_INAA            EC               Cr            Cr_INAA     
#>  Min.   : 3.00   Min.   :0.1400   Min.   :  2.20   Min.   : 11.0  
#>  1st Qu.:10.00   1st Qu.:0.2500   1st Qu.: 18.50   1st Qu.: 68.0  
#>  Median :13.00   Median :0.3100   Median : 28.35   Median : 99.0  
#>  Mean   :14.29   Mean   :0.3556   Mean   : 36.16   Mean   :116.2  
#>  3rd Qu.:17.00   3rd Qu.:0.3900   3rd Qu.: 41.60   3rd Qu.:130.0  
#>  Max.   :57.00   Max.   :4.2400   Max.   :471.00   Max.   :910.0  
#>  NA's   :1                                                        
#>     Cs_INAA            Cu            Eu_INAA            F_IC         
#>  Min.   :1.000   Min.   :  2.00   Min.   : 0.300   Min.   :0.000500  
#>  1st Qu.:1.000   1st Qu.: 10.32   1st Qu.: 0.800   1st Qu.:0.000500  
#>  Median :2.000   Median : 16.20   Median : 1.050   Median :0.000500  
#>  Mean   :1.953   Mean   : 21.96   Mean   : 1.239   Mean   :0.000553  
#>  3rd Qu.:2.000   3rd Qu.: 28.00   3rd Qu.: 1.400   3rd Qu.:0.000500  
#>  Max.   :8.000   Max.   :149.00   Max.   :14.300   Max.   :0.004100  
#>  NA's   :372                                                         
#>        Fe           Fe_INAA           Fe2O3           Hf_INAA      
#>  Min.   : 3310   Min.   :  6800   Min.   : 0.850   Min.   :  2.00  
#>  1st Qu.:10500   1st Qu.: 27025   1st Qu.: 3.730   1st Qu.:  4.00  
#>  Median :14700   Median : 35750   Median : 4.910   Median :  6.00  
#>  Mean   :17236   Mean   : 37815   Mean   : 5.161   Mean   :  6.47  
#>  3rd Qu.:20975   3rd Qu.: 46575   3rd Qu.: 6.367   3rd Qu.:  7.00  
#>  Max.   :79200   Max.   :119000   Max.   :17.660   Max.   :120.00  
#>                                                                    
#>        Hg            Hg_INAA       Ir_INAA          K              K2O       
#>  Min.   :0.0200   Min.   :1.0   Min.   : NA   Min.   :  200   Min.   :0.440  
#>  1st Qu.:0.0200   1st Qu.:1.0   1st Qu.: NA   1st Qu.:  700   1st Qu.:1.360  
#>  Median :0.0300   Median :1.0   Median : NA   Median : 1100   Median :1.710  
#>  Mean   :0.0289   Mean   :1.6   Mean   :NaN   Mean   : 1485   Mean   :1.883  
#>  3rd Qu.:0.0300   3rd Qu.:2.0   3rd Qu.: NA   3rd Qu.: 1800   3rd Qu.:2.268  
#>  Max.   :0.1600   Max.   :3.0   Max.   : NA   Max.   :11000   Max.   :6.320  
#>  NA's   :340      NA's   :601   NA's   :606   NA's   :3                      
#>        La            La_INAA             Li              LOI        
#>  Min.   :  3.50   Min.   :  6.10   Min.   : 1.700   Min.   : 0.340  
#>  1st Qu.:  9.60   1st Qu.: 17.00   1st Qu.: 4.700   1st Qu.: 1.190  
#>  Median : 12.80   Median : 24.00   Median : 7.200   Median : 1.800  
#>  Mean   : 17.94   Mean   : 30.64   Mean   : 9.129   Mean   : 2.337  
#>  3rd Qu.: 19.20   3rd Qu.: 35.75   3rd Qu.:11.100   3rd Qu.: 2.910  
#>  Max.   :203.00   Max.   :310.00   Max.   :70.900   Max.   :16.410  
#>                                                                     
#>     Lu_INAA          wt_INAA            Mg             MgO        
#>  Min.   :0.0500   Min.   :28.44   Min.   :  370   Min.   : 0.210  
#>  1st Qu.:0.2200   1st Qu.:31.83   1st Qu.: 2580   1st Qu.: 1.400  
#>  Median :0.3000   Median :33.58   Median : 3720   Median : 1.910  
#>  Mean   :0.3682   Mean   :33.75   Mean   : 4741   Mean   : 2.117  
#>  3rd Qu.:0.4500   3rd Qu.:35.41   3rd Qu.: 5368   3rd Qu.: 2.570  
#>  Max.   :2.6700   Max.   :39.70   Max.   :70500   Max.   :12.140  
#>                                                                   
#>        Mn              MnO               Mo            Mo_INAA      
#>  Min.   :  33.8   Min.   :0.0200   Min.   :0.2000   Min.   : 1.000  
#>  1st Qu.:  95.0   1st Qu.:0.0500   1st Qu.:0.3000   1st Qu.: 3.000  
#>  Median : 128.5   Median :0.0700   Median :0.4000   Median : 4.000  
#>  Mean   : 185.2   Mean   :0.0767   Mean   :0.5155   Mean   : 4.773  
#>  3rd Qu.: 201.5   3rd Qu.:0.0900   3rd Qu.:0.6000   3rd Qu.: 6.000  
#>  Max.   :2140.0   Max.   :0.4600   Max.   :3.4000   Max.   :19.000  
#>                                    NA's   :464      NA's   :509     
#>        Na             Na_INAA           Na2O          Nd_INAA      
#>  Min.   :   20.0   Min.   : 1900   Min.   :0.110   Min.   :  5.00  
#>  1st Qu.:   90.0   1st Qu.:21400   1st Qu.:2.590   1st Qu.: 13.00  
#>  Median :  140.0   Median :27600   Median :3.310   Median : 18.00  
#>  Mean   :  337.7   Mean   :25478   Mean   :3.053   Mean   : 22.62  
#>  3rd Qu.:  220.0   3rd Qu.:30600   3rd Qu.:3.660   3rd Qu.: 26.00  
#>  Max.   :19400.0   Max.   :53100   Max.   :6.570   Max.   :220.00  
#>                                                    NA's   :8       
#>        Ni            Ni_INAA          NO3_IC                P         
#>  Min.   :  1.20   Min.   : 45.0   Min.   :0.0005000   Min.   :  59.0  
#>  1st Qu.: 12.22   1st Qu.: 71.5   1st Qu.:0.0005000   1st Qu.: 281.2  
#>  Median : 18.65   Median : 90.0   Median :0.0005000   Median : 393.0  
#>  Mean   : 23.41   Mean   :101.6   Mean   :0.0006137   Mean   : 446.1  
#>  3rd Qu.: 28.27   3rd Qu.:110.0   3rd Qu.:0.0005000   3rd Qu.: 537.0  
#>  Max.   :228.00   Max.   :280.0   Max.   :0.0066000   Max.   :7170.0  
#>                   NA's   :523                                         
#>       P2O5             Pb               pH            PO4_IC        
#>  Min.   :0.010   Min.   : 0.300   Min.   :3.700   Min.   :0.002000  
#>  1st Qu.:0.070   1st Qu.: 1.200   1st Qu.:5.600   1st Qu.:0.002000  
#>  Median :0.090   Median : 1.600   Median :5.800   Median :0.002000  
#>  Mean   :0.105   Mean   : 2.748   Mean   :5.832   Mean   :0.002003  
#>  3rd Qu.:0.120   3rd Qu.: 3.100   3rd Qu.:6.000   3rd Qu.:0.002000  
#>  Max.   :1.350   Max.   :45.300   Max.   :7.600   Max.   :0.004000  
#>                                                                     
#>        Rb               S                Sb            Sb_INAA      
#>  Min.   : 16.00   Min.   :  7.00   Min.   :0.0100   Min.   :0.1000  
#>  1st Qu.: 41.00   1st Qu.: 20.00   1st Qu.:0.0150   1st Qu.:0.2000  
#>  Median : 55.50   Median : 30.00   Median :0.0230   Median :0.3000  
#>  Mean   : 63.57   Mean   : 40.84   Mean   :0.0498   Mean   :0.3887  
#>  3rd Qu.: 77.00   3rd Qu.: 45.50   3rd Qu.:0.0432   3rd Qu.:0.4000  
#>  Max.   :270.00   Max.   :531.00   Max.   :1.4200   Max.   :3.0000  
#>  NA's   :38       NA's   :3        NA's   :326      NA's   :455     
#>        Sc           Sc_INAA           Se             Se_INAA      
#>  Min.   : 0.20   Min.   : 1.7   Min.   :0.01000   Min.   : 3.000  
#>  1st Qu.: 1.70   1st Qu.: 9.2   1st Qu.:0.02800   1st Qu.: 3.250  
#>  Median : 2.30   Median :13.0   Median :0.04800   Median : 4.500  
#>  Mean   : 2.82   Mean   :13.6   Mean   :0.06947   Mean   : 6.167  
#>  3rd Qu.: 3.40   3rd Qu.:17.0   3rd Qu.:0.08200   3rd Qu.: 8.750  
#>  Max.   :15.40   Max.   :36.0   Max.   :1.15000   Max.   :12.000  
#>  NA's   :1                      NA's   :25        NA's   :600     
#>        Si             SiO2          Sm_INAA          Sn_INAA      
#>  Min.   : 50.0   Min.   :36.49   Min.   : 0.900   Min.   : 50.00  
#>  1st Qu.:110.0   1st Qu.:64.35   1st Qu.: 2.400   1st Qu.: 50.00  
#>  Median :140.0   Median :67.91   Median : 3.400   Median : 50.00  
#>  Mean   :154.2   Mean   :67.32   Mean   : 3.964   Mean   : 50.08  
#>  3rd Qu.:180.0   3rd Qu.:70.73   3rd Qu.: 4.675   3rd Qu.: 50.00  
#>  Max.   :590.0   Max.   :86.17   Max.   :37.000   Max.   :100.00  
#>                                                                   
#>      SO4_IC              Sr             Sr_INAA        SUM_XRF      
#>  Min.   :  1.000   Min.   :   1.60   Min.   : 500   Min.   : 97.70  
#>  1st Qu.:  3.000   1st Qu.:   5.50   1st Qu.: 525   1st Qu.: 99.21  
#>  Median :  4.000   Median :   7.70   Median : 650   Median : 99.60  
#>  Mean   :  5.546   Mean   :  25.34   Mean   : 689   Mean   : 99.53  
#>  3rd Qu.:  6.000   3rd Qu.:  11.95   3rd Qu.: 700   3rd Qu.: 99.88  
#>  Max.   :271.000   Max.   :1040.00   Max.   :1700   Max.   :101.89  
#>                                      NA's   :524                    
#>     Ta_INAA          Tb_INAA             Te               Th        
#>  Min.   : 0.500   Min.   :0.5000   Min.   :0.0030   Min.   : 3.000  
#>  1st Qu.: 0.900   1st Qu.:0.7000   1st Qu.:0.0060   1st Qu.: 5.000  
#>  Median : 1.200   Median :0.8000   Median :0.0100   Median : 7.000  
#>  Mean   : 2.786   Mean   :0.9898   Mean   :0.0132   Mean   : 8.358  
#>  3rd Qu.: 2.000   3rd Qu.:1.1000   3rd Qu.:0.0150   3rd Qu.:10.000  
#>  Max.   :41.000   Max.   :6.4000   Max.   :0.2710   Max.   :66.000  
#>  NA's   :429      NA's   :439      NA's   :137      NA's   :37      
#>     Th_INAA             Ti              TiO2           U_INAA     
#>  Min.   : 1.000   Min.   :  48.8   Min.   :0.090   Min.   : 0.70  
#>  1st Qu.: 3.900   1st Qu.: 560.0   1st Qu.:0.410   1st Qu.: 1.50  
#>  Median : 5.800   Median : 806.5   Median :0.580   Median : 1.95  
#>  Mean   : 7.164   Mean   : 895.5   Mean   :0.605   Mean   : 2.33  
#>  3rd Qu.: 9.400   3rd Qu.:1130.0   3rd Qu.:0.750   3rd Qu.: 2.60  
#>  Max.   :54.000   Max.   :5730.0   Max.   :3.170   Max.   :16.00  
#>                                                    NA's   :296    
#>        V              W_INAA          Y              Yb_INAA      
#>  Min.   :  4.50   Min.   : 1    Min.   :  0.900   Min.   : 0.300  
#>  1st Qu.: 21.55   1st Qu.: 3    1st Qu.:  3.100   1st Qu.: 1.400  
#>  Median : 30.90   Median : 4    Median :  4.400   Median : 1.900  
#>  Mean   : 34.99   Mean   : 4    Mean   :  6.366   Mean   : 2.377  
#>  3rd Qu.: 43.85   3rd Qu.: 5    3rd Qu.:  6.400   3rd Qu.: 3.000  
#>  Max.   :183.00   Max.   :10    Max.   :169.000   Max.   :19.900  
#>                   NA's   :593                                     
#>        Zn            Zn_INAA            ELEV           *COUN      
#>  Min.   :  3.70   Min.   : 50.00   Min.   : 20.0   Min.   :1.000  
#>  1st Qu.: 13.70   1st Qu.: 72.00   1st Qu.:140.0   1st Qu.:1.000  
#>  Median : 20.90   Median : 94.00   Median :200.0   Median :2.000  
#>  Mean   : 27.40   Mean   : 98.97   Mean   :205.6   Mean   :2.168  
#>  3rd Qu.: 33.73   3rd Qu.:118.00   3rd Qu.:260.0   3rd Qu.:3.000  
#>  Max.   :348.00   Max.   :356.00   Max.   :540.0   Max.   :3.000  
#>                   NA's   :237                                     
#>       *ASP            TOPC             LITO            Al_XRF     
#>  Min.   : 0.00   Min.   : 10.00   Min.   :  1.00   Min.   : 2.92  
#>  1st Qu.: 0.00   1st Qu.: 35.00   1st Qu.:  4.00   1st Qu.: 6.91  
#>  Median : 3.00   Median : 40.00   Median : 20.00   Median : 7.38  
#>  Mean   : 8.42   Mean   : 41.92   Mean   : 21.23   Mean   : 7.34  
#>  3rd Qu.:12.00   3rd Qu.: 49.00   3rd Qu.: 31.00   3rd Qu.: 7.82  
#>  Max.   :34.00   Max.   :123.00   Max.   :107.00   Max.   :12.08  
#>                                                                   
#>      Ca_XRF          Fe_XRF           K_XRF           Mg_XRF     
#>  Min.   :0.030   Min.   : 0.590   Min.   :0.360   Min.   :0.120  
#>  1st Qu.:1.603   1st Qu.: 2.600   1st Qu.:1.120   1st Qu.:0.840  
#>  Median :2.170   Median : 3.430   Median :1.410   Median :1.150  
#>  Mean   :2.133   Mean   : 3.605   Mean   :1.558   Mean   :1.271  
#>  3rd Qu.:2.690   3rd Qu.: 4.447   3rd Qu.:1.877   3rd Qu.:1.540  
#>  Max.   :6.760   Max.   :12.350   Max.   :5.240   Max.   :7.320  
#>                                                                  
#>      Mn_XRF            Na_XRF          P_XRF             Si_XRF     
#>  Min.   :0.01500   Min.   :0.080   Min.   :0.00400   Min.   :17.05  
#>  1st Qu.:0.03800   1st Qu.:1.917   1st Qu.:0.03000   1st Qu.:30.07  
#>  Median :0.05400   Median :2.450   Median :0.03900   Median :31.74  
#>  Mean   :0.05879   Mean   :2.260   Mean   :0.04536   Mean   :31.46  
#>  3rd Qu.:0.06900   3rd Qu.:2.710   3rd Qu.:0.05200   3rd Qu.:33.06  
#>  Max.   :0.35600   Max.   :4.870   Max.   :0.58900   Max.   :40.27  
#>                                                                     
#>      Ti_XRF     
#>  Min.   :0.053  
#>  1st Qu.:0.245  
#>  Median :0.347  
#>  Mean   :0.362  
#>  3rd Qu.:0.449  
#>  Max.   :1.900  
#>