Random Forest Imputation

Impute missing values based on a random forest model using ranger::ranger()

rangerImpute(
  formula,
  data,
  imp_var = TRUE,
  imp_suffix = "imp",
  ...,
  verbose = FALSE,
  median = FALSE
)

Arguments

formula
model formula for the imputation

data
A data.frame containing the data

imp_var
TRUE/FALSE if a TRUE/FALSE variables for each imputed variable should be created show the imputation status

imp_suffix
suffix used for TF imputation variables

...
Arguments passed to ranger::ranger()

verbose
Show the number of observations used for training and evaluating the RF-Model. This parameter is also passed down to ranger::ranger() to show computation status.

median
Use the median (rather than the arithmetic mean) to average the values of individual trees for a more robust estimate.

Value

the imputed data set.

See also

Examples

data(sleep)
rangerImpute(Dream+NonD~BodyWgt+BrainWgt,data=sleep)

#>     BodyWgt BrainWgt      NonD    Dream Sleep  Span  Gest Pred Exp Danger
#> 1  6654.000  5712.00  3.090897 1.473493   3.3  38.6 645.0    3   5      3
#> 2     1.000     6.60  6.300000 2.000000   8.3   4.5  42.0    3   1      3
#> 3     3.385    44.50  9.331343 2.641007  12.5  14.0  60.0    1   1      1
#> 4     0.920     5.70  8.421320 1.905647  16.5    NA  25.0    5   2      3
#> 5  2547.000  4603.00  2.100000 1.800000   3.9  69.0 624.0    3   5      4
#> 6    10.550   179.50  9.100000 0.700000   9.8  27.0 180.0    4   4      4
#> 7     0.023     0.30 15.800000 3.900000  19.7  19.0  35.0    1   1      1
#> 8   160.000   169.00  5.200000 1.000000   6.2  30.4 392.0    4   5      4
#> 9     3.300    25.60 10.900000 3.600000  14.5  28.0  63.0    1   2      1
#> 10   52.160   440.00  8.300000 1.400000   9.7  50.0 230.0    1   1      1
#> 11    0.425     6.40 11.000000 1.500000  12.5   7.0 112.0    5   4      4
#> 12  465.000   423.00  3.200000 0.700000   3.9  30.0 281.0    5   5      5
#> 13    0.550     2.40  7.600000 2.700000  10.3    NA    NA    2   1      2
#> 14  187.100   419.00  5.298583 1.349243   3.1  40.0 365.0    5   5      5
#> 15    0.075     1.20  6.300000 2.100000   8.4   3.5  42.0    1   1      1
#> 16    3.000    25.00  8.600000 0.000000   8.6  50.0  28.0    2   2      2
#> 17    0.785     3.50  6.600000 4.100000  10.7   6.0  42.0    2   2      2
#> 18    0.200     5.00  9.500000 1.200000  10.7  10.4 120.0    2   2      2
#> 19    1.410    17.50  4.800000 1.300000   6.1  34.0    NA    1   2      1
#> 20   60.000    81.00 12.000000 6.100000  18.1   7.0    NA    1   1      1
#> 21  529.000   680.00  3.451557 0.300000    NA  28.0 400.0    5   5      5
#> 22   27.660   115.00  3.300000 0.500000   3.8  20.0 148.0    5   5      5
#> 23    0.120     1.00 11.000000 3.400000  14.4   3.9  16.0    3   1      2
#> 24  207.000   406.00  5.332770 1.366623  12.0  39.3 252.0    1   4      1
#> 25   85.000   325.00  4.700000 1.500000   6.2  41.0 310.0    1   3      1
#> 26   36.330   119.50  6.223953 1.036677  13.0  16.2  63.0    1   1      1
#> 27    0.101     4.00 10.400000 3.400000  13.8   9.0  28.0    5   1      3
#> 28    1.040     5.50  7.400000 0.800000   8.2   7.6  68.0    5   3      4
#> 29  521.000   655.00  2.100000 0.800000   2.9  46.0 336.0    5   5      5
#> 30  100.000   157.00  5.228487 1.320483  10.8  22.4 100.0    1   1      1
#> 31   35.000    56.00  8.054250 2.211603    NA  16.3  33.0    3   5      4
#> 32    0.005     0.14  7.700000 1.400000   9.1   2.6  21.5    5   2      4
#> 33    0.010     0.25 17.900000 2.000000  19.9  24.0  50.0    1   1      1
#> 34   62.000  1320.00  6.100000 1.900000   8.0 100.0 267.0    1   1      1
#> 35    0.122     3.00  8.200000 2.400000  10.6    NA  30.0    2   1      1
#> 36    1.350     8.10  8.400000 2.800000  11.2    NA  45.0    3   1      3
#> 37    0.023     0.40 11.900000 1.300000  13.2   3.2  19.0    4   1      3
#> 38    0.048     0.33 10.800000 2.000000  12.8   2.0  30.0    4   1      3
#> 39    1.700     6.30 13.800000 5.600000  19.4   5.0  12.0    2   1      1
#> 40    3.500    10.80 14.300000 3.100000  17.4   6.5 120.0    2   1      1
#> 41  250.000   490.00  5.419813 1.000000    NA  23.6 440.0    5   5      5
#> 42    0.480    15.50 15.200000 1.800000  17.0  12.0 140.0    2   2      2
#> 43   10.000   115.00 10.000000 0.900000  10.9  20.2 170.0    4   4      4
#> 44    1.620    11.40 11.900000 1.800000  13.7  13.0  17.0    2   1      2
#> 45  192.000   180.00  6.500000 1.900000   8.4  27.0 115.0    4   4      4
#> 46    2.500    12.10  7.500000 0.900000   8.4  18.0  31.0    5   5      5
#> 47    4.288    39.20  8.534407 2.121473  12.5  13.7  63.0    2   2      2
#> 48    0.280     1.90 10.600000 2.600000  13.2   4.7  21.0    3   1      3
#> 49    4.235    50.40  7.400000 2.400000   9.8   9.8  52.0    1   1      1
#> 50    6.800   179.00  8.400000 1.200000   9.6  29.0 164.0    2   3      2
#> 51    0.750    12.30  5.700000 0.900000   6.6   7.0 225.0    2   2      2
#> 52    3.600    21.00  4.900000 0.500000   5.4   6.0 225.0    3   2      3
#> 53   14.830    98.20  8.652533 1.003117   2.6  17.0 150.0    5   5      5
#> 54   55.500   175.00  3.200000 0.600000   3.8  20.0 151.0    5   5      5
#> 55    1.400    12.50  6.407853 1.323253  11.0  12.7  90.0    2   2      2
#> 56    0.060     1.00  8.100000 2.200000  10.3   3.5    NA    3   1      2
#> 57    0.900     2.60 11.000000 2.300000  13.3   4.5  60.0    2   1      2
#> 58    2.000    12.30  4.900000 0.500000   5.4   7.5 200.0    3   1      3
#> 59    0.104     2.50 13.200000 2.600000  15.8   2.3  46.0    3   2      2
#> 60    4.190    58.00  9.700000 0.600000  10.3  24.0 210.0    4   3      4
#> 61    3.500     3.90 12.800000 6.600000  19.4   3.0  14.0    2   1      1
#> 62    4.050    17.00  7.418381 1.354740    NA  13.0  38.0    3   1      1
#>    Dream_imp NonD_imp
#> 1       TRUE     TRUE
#> 2      FALSE    FALSE
#> 3       TRUE     TRUE
#> 4       TRUE     TRUE
#> 5      FALSE    FALSE
#> 6      FALSE    FALSE
#> 7      FALSE    FALSE
#> 8      FALSE    FALSE
#> 9      FALSE    FALSE
#> 10     FALSE    FALSE
#> 11     FALSE    FALSE
#> 12     FALSE    FALSE
#> 13     FALSE    FALSE
#> 14      TRUE     TRUE
#> 15     FALSE    FALSE
#> 16     FALSE    FALSE
#> 17     FALSE    FALSE
#> 18     FALSE    FALSE
#> 19     FALSE    FALSE
#> 20     FALSE    FALSE
#> 21     FALSE     TRUE
#> 22     FALSE    FALSE
#> 23     FALSE    FALSE
#> 24      TRUE     TRUE
#> 25     FALSE    FALSE
#> 26      TRUE     TRUE
#> 27     FALSE    FALSE
#> 28     FALSE    FALSE
#> 29     FALSE    FALSE
#> 30      TRUE     TRUE
#> 31      TRUE     TRUE
#> 32     FALSE    FALSE
#> 33     FALSE    FALSE
#> 34     FALSE    FALSE
#> 35     FALSE    FALSE
#> 36     FALSE    FALSE
#> 37     FALSE    FALSE
#> 38     FALSE    FALSE
#> 39     FALSE    FALSE
#> 40     FALSE    FALSE
#> 41     FALSE     TRUE
#> 42     FALSE    FALSE
#> 43     FALSE    FALSE
#> 44     FALSE    FALSE
#> 45     FALSE    FALSE
#> 46     FALSE    FALSE
#> 47      TRUE     TRUE
#> 48     FALSE    FALSE
#> 49     FALSE    FALSE
#> 50     FALSE    FALSE
#> 51     FALSE    FALSE
#> 52     FALSE    FALSE
#> 53      TRUE     TRUE
#> 54     FALSE    FALSE
#> 55      TRUE     TRUE
#> 56     FALSE    FALSE
#> 57     FALSE    FALSE
#> 58     FALSE    FALSE
#> 59     FALSE    FALSE
#> 60     FALSE    FALSE
#> 61     FALSE    FALSE
#> 62      TRUE     TRUE